First commit of showdetected
showdetected:

Provide minimal trilateralization (location) services and recommend clients to be "kicked" off of their currently associated AP in order to force it to roam to an AP with a better signal to noise ratio. (a closer AP)

note:
Contains a hard-coded values for "CONTROLLER" and "SNMPCOMMUNITY".
Contains a (commented) "ruckusconf" line that aquires required data.
That "ruckusconf" line will need to be run manually before this script.
This script takes a while to run, now that I've got over 100 APs.
This script creates a "getdetected" working file, which may contain useful (location) info but is easier to read when filtered through the "./apnames.sed" file.